Fix movement speed bug
Finally fixed the bug which was preventing the player from moving at
full speed. The problem was that the players position was being updated
to the server's player position every tick. This sometimes caused the
player to move to an old position which would subsequently cause the
player to move too far in one tick and disconnect.