fix(client): fix reconnect handler host issue
The reconnect packet handler would default to using the nexus server's
IP if the host in the reconnect packet was null. This is a problem when
connecting to dungeons, because they use the same host as the current
realm, so the reconnect packet host is null. Now, the host is left
untouched if the reconnect packet has no host.

Closes #26