feat(build): remove gulp
This commit removes gulp and its dependencies. The only reason gulp was
used for building was to copy .json files across into the dist/
directory in case there were any in the plugins directory. However, the
same thing can be achieved by just enabling the "resolveJsonModule" in
the tsconfig, so gulp is no longer needed.