fix(client): fix design issue with movement
This commit fixes a design issue with the client's movement system.
Previously, the position was only updated in the newtick handler when
the client had to reply with a move packet. This approach would
sometimes cause the client to "jump" over enemy projectiles, which in
turn would cause a disconnect because the server never receives the
appropriate PlayerHitPacket.
The design has been changed so that the position is now updated in the
frameUpdateTimer, which runs roughly 30 times per second.
The code for projectile checking is now also run in the frameUpdateTimer
instead of using its own timer.