feat(acc-service): make AccountService env based
The AccountService is now an instance based class instead of a static
one, and takes an environment as a constructor parameter.
Better caching capabilities have also been added. The account service
will now cache servers and character info, and will prefer the cache
over web requests when fetching the info.