feat(runtime): add runtime
This commit adds the Runtime class to nrelay. The runtime is a new
concept for nrelay which is designed to change the way nrelay is used.
At the moment, you have to download the source, add your plugins, and
build it yourself. This is great if you want to hack on the source, but
terrible if you just want to make a quick new project and start writing
plugins. The runtime can be created and pointed to a folder which
represents a "project". The project folder is expected to have various
files in it suchs as the accounts.json, versions.json, etc, and the
plugin files. This means that nrelay can be installed as an npm
dependency into a standalone project, and then a runtime can just be
created from that project.