feat(runtime): add plugin loading to runtime
This commit changes the LibraryManager to be instance based, and adds an
instance of it to the runtime class. This is the final change in
converting all of the services to the new "runtime" architecture. Some
changes were also made to the packet and library hooks in order to store
the information there until the runtime can load it.