fix(plugins): fix valid packet hooks being ignored
Packet hooks are ignored if the packet they are hooking does not appear
in the VALID_PACKET_HOOKS constant. However, this list was incomplete
which meant that some packet hooks that were valid would be ignored. The
list has been changed to use the exports from the incoming packets
folder, so it will always be complete.