feat(net): use @realmlib/net for networking
This commit switches to the @realmlib/net package for networking related
code. This package is re-exported from nrelay, so packets can still be
imported from the nrelay namespace as to not break plugins.
Some parser functions have been added which are used to parse object
status data. Previously, these were static functions on the
ObjectStatusData class, but these methods do not exist in @realmlib/net.
A new util has also been added to create connections that may or may not
be routed through a proxy in a more clean way.