feat(args): remove args parser
The args parser has been removed in favor of using Yargs to provide
args. The args object is provided to a runtime instance when it is run.
A new method has also been added to the runtime for updating the
buildversion. This could be done from other methods available on the
runtime class, but this way is more convenient.