feat(client): improve death handler and auto nexus
This commit adds proper death handling to the client class. When a death
occurs, the current character id is set to the value of the next
character id, and the next character id is incremented. Some handlers
have also been added for failure messages related to loading a dead
character, or loading a character that was not found.
The auto nexus has also been simplified, and in doing so, has been
improved.