fix(client): fix enemy hit mechanics
This commit fixes several problems with the code involved in generating
enemy hit packets.

+ Packets will no longer be sent for an enemy which has
died.
+ The order of operations when updating entities has been corrected.
+ Invincible enemies are not hit.

This commit also fixes a race condition which would result in a crash
when autonexusing with projectiles still on the screen.

It also adds several debug log messages, and a regular log message which
logs the client's health when it gets hit.