chore(client): improve connection logic
This commit makes several changes to the connection logic for the
client. Most notably, the client is not taken back to the nexus when the
connection is closed. The connection scheduler is also now used in the
client to handle reconnect delays. The client will also now handle
account in use errors, and will try to reconnect if it fails to connect.