feat(hooks): make hooks less restrictive
This commit changes how packet hooks are loaded and called in order to
make the function signature requirements less strict. The only
requirement now is that a packet appears at least somewhere in the
method's type signature. A client parameter is not required anymore.
Parameters which are not of the type Client or Packet will raise a
warning, but will not result in an error.